pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoannotate-vision"
version = "0.1.0"
description = "SOTA unsupervised auto-annotation SDK for image classification"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "MetaMind Innovations", email = "research@metamind-innovations.com"}
]
keywords = [
    "image-annotation",
    "clustering", 
    "computer-vision",
    "auto-labeling",
    "machine-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.0.0",
    "umap-learn>=0.5.3",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]
hdbscan = [
    "hdbscan>=0.8.29",
]

[project.scripts]
autoannotate = "autoannotate.ui.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/autoannotate-vision"
Documentation = "https://autoannotate-vision.readthedocs.io"
Repository = "https://github.com/yourusername/autoannotate-vision"
"Bug Tracker" = "https://github.com/yourusername/autoannotate-vision/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
autoannotate = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=autoannotate --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true